Added code to train models with joint transition params on multiple datasets. Specified model takes priority over load_model now. Updated state selection cutoff to 0.01
