+++ cy2path/__init__.py
@@ -1,7 +1,7 @@
 from .sampling import sample_state_probability, iterate_state_probability
 from .simulation import sample_markov_chains, iterate_markov_chain
 from .pseudotime import infer_pseudotime, add_expected_steps
-from .dynamics import infer_latent_dynamics, infer_kinetic_clusters, infer_lineage_probabilities, infer_latent_paths
+from .dynamics import infer_dynamics, infer_joint_dynamics, infer_kinetic_clusters, infer_lineage_probabilities, infer_latent_paths
 from .selection import run_search
 
 from .cytopath import infer_cytopath_lineages
